# Conflicts:
#	src/main/kotlin/ar/edu/unsam/phm/bootstrap/BookReservationInitializer.kt
#	src/main/kotlin/ar/edu/unsam/phm/domain/Book.kt
#	src/main/kotlin/ar/edu/unsam/phm/domain/BookTypes.kt
#	src/main/kotlin/ar/edu/unsam/phm/dto/BookDTO.kt
#	src/main/kotlin/ar/edu/unsam/phm/repository/CrudBookRepository.kt
#	src/main/kotlin/ar/edu/unsam/phm/services/BookService.kt
#	src/main/kotlin/ar/edu/unsam/phm/services/ReservationService.kt

+++ src/main/kotlin/ar/edu/unsam/phm/dto/BookDTO.kt
@@ -20,7 +20,7 @@ data class BookDTO(
     var condition: String,
     var owner: UserDTO,
     var imageSrc: String,
-    var bookBibliokarmas: Int = 0,
+    var bookBibliokarmas: Long = 0,
     var rating: Double = 0.0
 )
